pyproject.toml
[project]
name = "writing-team"
version = "2.0.0"
description = "Multi-agent AI book writing system powered by Google ADK and Gemini"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "pyzotero>=1.5.0",
    "PyMuPDF>=1.25.0",
    "python-docx>=1.1.0",
    "textstat>=0.7.0",
    "httpx>=0.28.0",
    "aiofiles>=24.0.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agents", "tools"]

